controller/signup.py
from flask import Blueprint, render_template, request, redirect
from module.githubApi import GithubApi
from module.InMemoryCache import inMemoryCacheInstance
import hashlib
import jwt
from repository.repositoryProfile import profile_repository
from repository.repositoryProfile import token_repository
from repository.repositoryProfile import TokenTable
from repository.repositoryProfile import UserTable
from datetime import datetime

from configuration.config import Config

from batch.CommitCountScheduler import CommitCountScheduler


bp = Blueprint('signup', __name__)

githubApi = GithubApi()

@bp.route("/signup/redirect", methods=['GET'])
def singupRedirect():
    githubLoginUrl = githubApi.getLoginUrl()
    return redirect(githubLoginUrl)

@bp.route("/signup/complete", methods=["GET"])
def signupComplete():
    
    code = request.args.get('code')
    
    if code:
        accessToken = githubApi.getAccessToken(code)

        # GitHub로부터 받은 access token을 세션에 저장하거나, 필요한 처리를 합니다.
        key = hashlib.sha256(accessToken.encode()).hexdigest()
        print("🍎" + key)
        inMemoryCacheInstance.set(key, accessToken)
        # 이후 signup.html 페이지로 리다이렉트합니다.
        return redirect(f'/signup?code={key}')
    else:
        return "GitHub 인증 실패", 400


@bp.route("/signup", methods=['GET'])
def signup():
    # 사용자가 처음으로 접근하면 GitHub 로그인 페이지로 리다이렉트
    code = request.args.get('code')
    print("🍎🍎" + code)
    return render_template('signup.html',code=code)
       
        
@bp.route("/signup/update", methods=['POST'])
def signupUpdate():

    # TODO: 중복 검사 해야 함. 

    code = request.args.get('code')
    github_access_token = inMemoryCacheInstance.get(code)
    print("🍎🍎" + code)

    id = request.form['id']
    password = request.form['password']
    passwordconfirm = request.form['password-confirm']
    cardinal = request.form['cardinal']
    number = request.form['number']
    intro = request.form['intro']

    githubUserInfo = githubApi.getUserInfo(githubAccessToken=github_access_token)

    pic_url = ""
    name = ""
    git = ""
    gitId = ""
    if githubUserInfo is not None:
        pic_url = githubUserInfo['avatar_url']
        name = githubUserInfo['name']
        git = githubUserInfo['html_url']
        gitId = githubUserInfo['login']

        if profile_repository.read_git(git) == git:
            # git 로 검색해서 있으면 중복 처리
            return redirect(f'/signup/fail?message=aleady')

    # 입력받은 데이터를 usertable DB에 저장
    usertable = UserTable(
        _id=None,  # MongoDB에서 자동 생성되므로 None으로 설정
        id = id,
        password = password,
        pic_url=pic_url,
        generation=cardinal,
        num=number,
        name=name, 
        like=0,
        git=git, 
        gitId=gitId,
        bio=intro,
        githubaccesstoken = github_access_token
    )
    created_user = profile_repository.create(usertable) 
    user_id = created_user._id
    print(f"생성된 유저의 _id: {user_id}")
 
    # jwt 토큰 발급
    payload = {"userId": user_id}

    config =Config()
    secret_key = config.find("MY_SECRET_KEY")
    algo = config.find("SECRET_KEY_ALGO")
    accessToken = jwt.encode(payload, secret_key, algorithm=algo)
    refreshToken = jwt.encode(payload, secret_key, algorithm=algo)

    # main 화면에 전달 
    key = hashlib.sha256(accessToken.encode()).hexdigest()
    inMemoryCacheInstance.set(key, accessToken)
 
    # tokentable에 token정보 추가
    tokentable = TokenTable(
        userId = user_id,
        accesstoken = accessToken,
        refreshtoken = refreshToken,
        updateat = None,
        createdat = datetime.now()
    )
    created_token = token_repository.create(tokentable) 
    print(f"생성된 유저의 userId: {created_token.userId}")    

    # count batch refresh 
    # CommitCountScheduler().job()

    return redirect('/main')


@bp.route("/signup/fail")
def sigonupFail():
    message = request.args.get('message')
    return render_template('signupFail.html', message="이미 가입된 아이디입니다.")
