requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
frozenlist==1.4.1
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
PyJWT==2.9.0
pymongo==4.8.0
pytz==2024.1
requests==2.32.3
six==1.16.0
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.8
zipp==3.20.1
